+++ packages/transcription-local/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mcut/transcription-local
 
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mcut/transcription@0.1.0-alpha.12
+
 ## 1.0.0-alpha.14
 
 ### Patch Changes